+++ src/main/java/ru/kata/spring/boot_security/demo/repository/UserDao.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface UserDao {
     void saveUser(User user);
-    User getUserById(int id);
+    User getUserById(Long id);
     void updateUser(User user);
-    void deleteUserById(int id);
+    void deleteUserById(Long id);
     List<User> getAllUsers();
     User findByUsername(String name);
 }
