+++ src/main/java/ru/kata/spring/boot_security/demo/service/UserServiceImpl.java
@@ -27,7 +27,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getUserById(int id) {
+    public User getUserById(Long id) {
         return this.userDao.getUserById(id);
     }
 
@@ -39,7 +39,7 @@ public class UserServiceImpl implements UserService {
 
     @Transactional
     @Override
-    public void deleteUserById(int id) {
+    public void deleteUserById(Long id) {
         this.userDao.deleteUserById(id);
     }
